Move insertion point types under `render`
Also:
 * Move `FilePattern` in a separate file.
 * Rename `actions.proto` to `render_actions.proto` for a better reading experience for new developers.
