+++ api/src/main/kotlin/io/spine/protodata/render/InsertionPointPrinter.kt
@@ -27,7 +27,7 @@
 package io.spine.protodata.render
 
 import io.spine.core.userId
-import io.spine.protodata.event.insertionPointPrinted
+import io.spine.protodata.render.event.insertionPointPrinted
 import io.spine.protodata.file
 import io.spine.protodata.render.CoordinatesFactory.Companion.endOfFile
 import io.spine.text.TextCoordinates
