+++ api/src/main/proto/spine/protodata/events.proto
@@ -397,26 +397,3 @@ message DependencyDiscovered {
     // The source file of the dependency.
     ProtobufSourceFile source = 2 [(required) = true];
 }
-
-// Emitted when an insertion point printer adds an insertion point to a code file.
-message InsertionPointPrinted {
-
-    // Path to the code file.
-    //
-    // This path is relative to the source file set which contains the file.
-    //
-    File file = 1 [(required) = true];
-
-    // The insertion point label.
-    //
-    // This is a unique text that allows ProtoData to find insertion points in the text later.
-    //
-    string label = 2 [(required) = true];
-
-    // The appearance of the insertion point in the code.
-    //
-    // This is typically a comment which contains the insertion point label and possibly some other
-    // syntax. New code can be inserted right after this substring in the code file.
-    //
-    string representation_in_code = 3 [(required) = true];
-}
