+++ api/src/main/proto/spine/protodata/file_pattern.proto
@@ -31,39 +31,23 @@ package spine.protodata;
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.protodata.render";
-option java_outer_classname = "InsertionProto";
+option java_package = "io.spine.protodata";
+option java_outer_classname = "FilePatternProto";
 option java_multiple_files = true;
 
-import "spine/protodata/file.proto";
+// A pattern matching file paths.
+message FilePattern {
 
-// An insertion point added to a file.
-message InsertedPoint {
+    oneof kind {
+        option (is_required) = true;
 
-    // The relative path to the code file.
-    File file = 1 [(required) = true];
+        // The path must end with this string.
+        string suffix = 1;
 
-    // The label of the insertion point.
-    string label = 2 [(required) = true, (column) = true];
+        // The path must start with this string.
+        string prefix = 2;
 
-    // The code representation of the insertion point.
-    //
-    // Typically, this is a comment containing the insertion point label.
-    //
-    string representation_in_code = 3 [(required) = true];
-}
-
-// Insertion points added to a single code file.
-message InsertedPoints {
-    option (entity).kind = VIEW;
-
-    // The relative path to the code file.
-    File file = 1 [(required) = true];
-
-    // Insertion points added to the given file.
-    //
-    // Since a code file might have more than one instance of the same insertion point, this list
-    // may contain repetitions.
-    //
-    repeated InsertedPoint point = 2 [(validate) = true];
+        // The path must match this regular expression.
+        string regex = 3;
+    }
 }
