+++ backend/src/main/kotlin/io/spine/protodata/backend/InsertionPointsView.kt
@@ -29,10 +29,10 @@ package io.spine.protodata.backend
 import io.spine.core.External
 import io.spine.core.Subscribe
 import io.spine.protodata.File
-import io.spine.protodata.InsertedPoints
-import io.spine.protodata.event.InsertionPointPrinted
-import io.spine.protodata.insertedPoint
 import io.spine.protodata.plugin.View
+import io.spine.protodata.render.InsertedPoints
+import io.spine.protodata.render.insertedPoint
+import io.spine.protodata.render.event.InsertionPointPrinted
 import io.spine.server.entity.alter
 
 /**
